PYIC-1718: Rename issued-credentials lambda
