+++ lambdas/build-debug-credential-data/src/main/resources/IpvLambdaJsonLayout.json
