+++ lambdas/build-debug-credential-data/src/main/resources/log4j2.yaml
