+++ settings.gradle
@@ -6,7 +6,7 @@ include "lib",
 		"lambdas:retrieve-cri-oauth-access-token",
 		"lambdas:get-credential-issuer-config",
 		"lambdas:evaluate-gpg45-scores",
-		"lambdas:issuedcredentials",
+		"lambdas:build-debug-credential-data",
 		"lambdas:sessionstart",
 		"lambdas:useridentity",
 		"lambdas:process-journey-step",
